EdtitorService분리
