+++ src/main/java/com/green/vrink/user/service/EditorServiceImpl.java
@@ -1,0 +1,33 @@
+package com.green.vrink.user.service;
+
+import org.springframework.stereotype.Service;
+import org.springframework.transaction.annotation.Transactional;
+
+import com.green.vrink.user.dto.ApprovalDTO;
+import com.green.vrink.user.dto.EditorWriteDTO;
+import com.green.vrink.user.repository.interfaces.UserRepository;
+
+import lombok.RequiredArgsConstructor;
+import lombok.extern.slf4j.Slf4j;
+
+@Service
+@RequiredArgsConstructor
+@Slf4j
+public class EditorServiceImpl implements EditorService{
+	private final UserRepository userRepository;
+
+
+    @Transactional
+    public Integer requestApproval(ApprovalDTO approvalDTO) {
+
+
+
+
+        return userRepository.requestApproval(approvalDTO);
+    }
+    
+    @Transactional
+    public Integer requestEditorWrite(EditorWriteDTO editorWriteDTO) {
+    	return userRepository.editorWrite(editorWriteDTO);
+    }
+}
